+++ static_lib/header.h
@@ -4,7 +4,7 @@
 long int add(long int a, long int b);
 long int diff(long int a, long int b);
 long int prod(long int a, long int b);
-long int fract(long int a, long int b);
+double fract(double a, double b);
 unsigned long int fact(unsigned int n);
 double sqroot(double a);
